Update queries & structure
