+++ packages/shopify/src/api/utils/get-categories.ts
@@ -1,8 +1,7 @@
-import type { Category } from '../types/site'
-import { ShopifyConfig } from '../api'
-import { CollectionEdge } from '../../schema'
-import { normalizeCategory } from './normalize'
-import { getSiteCollectionsQuery } from './queries/get-all-collections-query'
+import { ShopifyConfig } from '..'
+import type { Category } from '../../types/site'
+import { CollectionEdge } from '../../../schema'
+import { normalizeCategory, getSiteCollectionsQuery } from '../../utils'
 
 export const getCategories = async ({
   fetch,
