+++ packages/shopify/src/utils/fragments/cart-details-fragment.ts
@@ -1,0 +1,82 @@
+export const cartDetailsFragment = /* GraphQL */ `
+  fragment cartDetails on Cart {
+    id
+    checkoutUrl
+    createdAt
+    updatedAt
+    lines(first: 10) {
+      edges {
+        node {
+          id
+          quantity
+          merchandise {
+            ... on ProductVariant {
+              id
+              id
+              sku
+              title
+              selectedOptions {
+                name
+                value
+              }
+              image {
+                url
+                altText
+                width
+                height
+              }
+              priceV2 {
+                amount
+                currencyCode
+              }
+              compareAtPriceV2 {
+                amount
+                currencyCode
+              }
+              product {
+                title
+                handle
+              }
+            }
+          }
+        }
+      }
+    }
+    attributes {
+      key
+      value
+    }
+    buyerIdentity {
+      email
+      customer {
+        id
+      }
+    }
+    estimatedCost {
+      totalAmount {
+        amount
+        currencyCode
+      }
+      subtotalAmount {
+        amount
+        currencyCode
+      }
+      totalTaxAmount {
+        amount
+        currencyCode
+      }
+      totalDutyAmount {
+        amount
+        currencyCode
+      }
+    }
+  }
+`
+
+export const userErrorsFragment = /* GraphQL */ `
+  fragment userErrors on UserError {
+    code
+    field
+    message
+  }
+`
