+++ packages/shopify/src/utils/fragments/customer-fragments.ts
@@ -1,0 +1,15 @@
+export const customerUserErrorsFragment = /* GraphQL */ `
+  fragment customerUserErrors on CustomerUserErrors {
+    code
+    field
+    message
+  }
+`
+
+export const customerAccessTokenFragment = /* GraphQL */ `
+  fragment customerAccessToken on CustomerAccessToken {
+    code
+    field
+    message
+  }
+`
