+++ packages/shopify/src/utils/handle-login.ts
@@ -1,6 +1,6 @@
 import { FetcherOptions } from '@vercel/commerce/utils/types'
 import { CustomerAccessTokenCreateInput } from '../../schema'
-import { setCustomerToken } from './customer-token'
+import { setCustomerToken } from './helpers'
 import { customerAccessTokenCreateMutation } from './mutations'
 import { throwUserErrors } from './throw-user-errors'
