+++ packages/shopify/src/utils/index.ts
@@ -1,15 +1,10 @@
+export { throwUserErrors } from './throw-user-errors'
 export { handleFetchResponse } from './handle-fetch-response'
-export { getSearchVariables } from './get-search-variables'
-export { getSortVariables } from './get-sort-variables'
-export { getBrands } from './get-brands'
-export { getCategories } from './get-categories'
-export { getCartId } from './get-cart-id'
 export { cartCreate } from './cart-create'
 export { handleLogin, handleAutomaticLogin } from './handle-login'
 export { handleAccountActivation } from './handle-account-activation'
-export { throwUserErrors } from './throw-user-errors'
 
+export * from './helpers'
 export * from './queries'
 export * from './mutations'
 export * from './normalize'
-export * from './customer-token'
