--- packages/shopify/src/utils/mutations/associate-customer-with-checkout.ts
@@ -1,17 +1,0 @@
-export const associateCustomerWithCheckoutMutation = /* GraphQl */ `
-mutation associateCustomerWithCheckout($checkoutId: ID!, $customerAccessToken: String!) {
-    checkoutCustomerAssociateV2(checkoutId: $checkoutId, customerAccessToken: $customerAccessToken) {
-      checkout {
-        id
-      }
-      checkoutUserErrors {
-        code
-        field
-        message
-      }
-      customer {
-        id
-      }
-    }
-  }
-`
