--- packages/shopify/src/utils/mutations/cart-create.ts
@@ -1,17 +1,0 @@
-import { cartDetailsFragment } from '../queries/get-cart-query'
-
-export const cartCreateMutation = /* GraphQL */ `
-  mutation cartCreate($input: CartInput = {}) {
-    cartCreate(input: $input) {
-      cart {
-        ...cartDetails
-      }
-      userErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-  ${cartDetailsFragment}
-`
