--- packages/shopify/src/utils/mutations/cart-lines-item-update.ts
@@ -1,17 +1,0 @@
-import { cartDetailsFragment } from '../queries/get-cart-query'
-
-export const cartLinesUpdateMutation = /* GraphQL */ `
-  mutation cartLinesUpdate($cartId: ID!, $lines: [CartLineUpdateInput!]!) {
-    cartLinesUpdate(cartId: $cartId, lines: $lines) {
-      cart {
-        ...cartDetails
-      }
-      userErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-  ${cartDetailsFragment}
-`
