--- packages/shopify/src/utils/mutations/customer-access-token-delete.ts
@@ -1,12 +1,0 @@
-export const customerAccessTokenDeleteMutation = /* GraphQL */ `
-  mutation customerAccessTokenDelete($customerAccessToken: String!) {
-    customerAccessTokenDelete(customerAccessToken: $customerAccessToken) {
-      deletedAccessToken
-      deletedCustomerAccessTokenId
-      userErrors {
-        field
-        message
-      }
-    }
-  }
-`
