--- packages/shopify/src/utils/mutations/customer-activate-by-url.ts
@@ -1,18 +1,0 @@
-export const customerActivateByUrlMutation = /* GraphQL */ `
-  mutation customerActivateByUrl($activationUrl: URL!, $password: String!) {
-    customerActivateByUrl(activationUrl: $activationUrl, password: $password) {
-      customer {
-        id
-      }
-      customerAccessToken {
-        accessToken
-        expiresAt
-      }
-      customerUserErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-`
