--- packages/shopify/src/utils/mutations/customer-activate.ts
@@ -1,18 +1,0 @@
-export const customerActivateMutation = /* GraphQL */ `
-  mutation customerActivate($id: ID!, $input: CustomerActivateInput!) {
-    customerActivate(id: $id, input: $input) {
-      customer {
-        id
-      }
-      customerAccessToken {
-        accessToken
-        expiresAt
-      }
-      customerUserErrors {
-        code
-        field
-        message
-      }
-    }
-  }
-`
