--- packages/shopify/src/utils/mutations/customer-create.ts
@@ -1,14 +1,0 @@
-export const customerCreateMutation = /* GraphQL */ `
-  mutation customerCreate($input: CustomerCreateInput!) {
-    customerCreate(input: $input) {
-      customerUserErrors {
-        code
-        field
-        message
-      }
-      customer {
-        id
-      }
-    }
-  }
-`
