+++ packages/shopify/src/utils/mutations/index.ts
@@ -1,9 +1,2 @@
-export { customerCreateMutation } from './customer-create'
-export { cartCreateMutation } from './cart-create'
-export { cartLinesAddMutation } from './cart-lines-item-add'
-export { cartLinesUpdateMutation } from './cart-lines-item-update'
-export { cartLinesRemoveMutation } from './cart-lines-item-remove'
-export { customerAccessTokenCreateMutation } from './customer-access-token-create'
-export { customerAccessTokenDeleteMutation } from './customer-access-token-delete'
-export { customerActivateMutation } from './customer-activate'
-export { customerActivateByUrlMutation } from './customer-activate-by-url'
+export * from './cart-mutations'
+export * from './customer-mutations'
