--- packages/shopify/src/utils/queries/get-site-info-query.ts
@@ -1,7 +1,0 @@
-export const getSiteInfoQuery = /* GraphQL */ `
-  query getSiteInfo {
-    shop {
-      name
-    }
-  }
-`
