+++ packages/shopify/src/utils/queries/index.ts
@@ -7,5 +7,4 @@ export { getCollectionProductsQuery } from './get-collection-products-query'
 export { getAllPagesQuery } from './get-all-pages-query'
 export { getPageQuery } from './get-page-query'
 export { getCustomerQuery } from './get-customer-query'
-export { getSiteInfoQuery } from './get-site-info-query'
 export { getCartQuery } from './get-cart-query'
